+++ src/main.js
@@ -47,6 +47,10 @@ const navigationConfig = goToScreen => {
     const characterSelect = data => goToScreen("character-select", data);
     const game = data => goToScreen("game", data);
     const results = data => goToScreen("results", data);
+    const achievements = data => { //eslint-disable-line
+        //TODO re-enable line below / remove eslint disable on line above once achievements screen is available
+        //goToScreen("achievements", data);
+    };
 
     return {
         loadscreen: {
@@ -59,6 +63,7 @@ const navigationConfig = goToScreen => {
             state: Home,
             routes: {
                 next: characterSelect,
+                achievements: achievements,
             },
         },
         "character-select": {
@@ -67,6 +72,7 @@ const navigationConfig = goToScreen => {
                 next: game,
                 home: home,
                 restart: home,
+                achievements: achievements,
             },
         },
         game: {
@@ -84,6 +90,7 @@ const navigationConfig = goToScreen => {
                 game: game,
                 restart: game,
                 home: home,
+                achievements: achievements,
             },
         },
     };