+++ test/components/select.spec.js
@@ -63,6 +63,7 @@ describe("Select Screen", () => {
                     characterSelect: {
                         choices: [{ asset: "character1" }, { asset: "character2" }, { asset: "character3" }],
                     },
+                    game: {},
                 },
             },
             qaMode: { active: false },
@@ -140,6 +141,20 @@ describe("Select Screen", () => {
             ]);
         });
     });
+
+    /*
+    // TODO the following line should be added back once the overlap issue is resolved NT:04:02:19
+    describe("achievements button", () => {
+        it("adds the achievement button when theme flag is set", () => {
+            selectScreen.context.config.theme.game.achievements = true;
+            selectScreen.create();
+
+            const actualButtons = addLayoutSpy.getCall(0).args[0];
+            const expectedButtons = ["home", "audio", "pauseNoReplay", "previous", "next", "continue", "achievements"];
+            assert.deepEqual(actualButtons, expectedButtons);
+        });
+    });
+    */
 
     describe("signals", () => {
         let signalSubscribeSpy;