+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 | Version | Description |
 |---------|-------------|
+| | Workaround for audio issue when tabbing away and pausing after music has finished. |
 | 1.0.4 | |
-| | Load global scripts synchronously in dev build |
+| | Load global scripts synchronously in dev build. |
 | 1.0.3 | |
 | | Lock to canvas only pending assessment of best practice for webgl performance. |
 | | Fix ios voiceover reading out zombie element . |